requirements.txt
babel==2.16.0
black==24.10.0
blinker==1.9.0
cachelib==0.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
docxcompose==1.4.0
Flask==3.1.0
Flask-Caching==2.3.0
gevent==24.11.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
jieba==0.42.1
Jinja2==3.1.5
langdetect==1.0.9
lxml==5.3.0
MarkupSafe==3.0.2
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
packaging==24.2
pathlib==1.0.1
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
prettytable==3.13.0
python-docx==1.1.2
qrcode==8.0
requests==2.32.3
six==1.17.0
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
